requirements.txt
torch==2.3.0+cu118
torchvision==0.18.0+cu118
torchaudio==2.3.0+cu118
faiss-gpu==1.7.4
sentence-transformers
langchain
transformers
scikit-learn
numpy
pandas
